rework to add explicit fourier sampler and generic mixture

+++ src/tree_world/models/fourier_metric.py
@@ -46,7 +46,12 @@ class FourierMetric(torch.nn.Module):
 
         # estimated displacements from thetas, made as small as possible solving across alphas -- but may not agree!
         # deltas has shape (..., J, d)
-        deltas = solve_for_deltas(delta_thetas.view(delta_thetas.shape[:-2] + (-1,)), self.K_dagger, self.lattice_basis, self.alphas)
+        deltas = solve_for_deltas(
+            delta_thetas.view(delta_thetas.shape[:-2] + (-1,)), 
+            self.K_dagger.to(delta_thetas.dtype), 
+            self.lattice_basis.to(delta_thetas.dtype), 
+            self.alphas.to(delta_thetas.dtype)
+        )
         mean_deltas = deltas.mean(dim=-2)
 
         return deltas, mean_deltas
